Implement automatic "future hooks"
Right now, Kirby only triggers the hook callbacks that have been registered before triggering the hook. In some cases (like "onload" events for complex plugins to notify other plugins) it makes sense to call a callback even if it is registered *after* triggering the hook.

This commit introduces a new parameter for the trigger method. If set to true, every newly registered hook with that name will get called with the arguments from the trigger call instantly after it has been registered. The behavior of calling hooks from the past is unchanged.