improved Kirbytag detection
Until now the Kirbytag detection was based on regex. Regex can not match
parentheses. Sometimes it is important to have nesting parentheses inside
a Kirbytag. This is especially true if you want to use mathematical
equations, for example using the library katex http://khan.github.io/KaTeX/.

This commit replaces the regular expression a by purely lexical processing
which allows nesting parentheses. Besides this new behaviour, it behaves
the same as before.

As a side effect it should be a lot of faster than a regex, but i have not
tested it for performance, so i can't be sure.