prev/next sortBy on multiple fields
Enabled the use of multiple sort fields vor prev/next. More or less related to https://github.com/getkirby/panel/pull/354.
I often use a date and time field for articles. When articles are posted on the same date, and only have a differen time stamp, this update enables the `prev` and `next` functions to return the correct previous and next article: `$article->prevVisible('date', 'asc', 'time', 'asc');`. Up till now I solved this with code in a controller but it would make more sense to be able to use the default prev/next functions.