Fix replacement of "placeholder" in strings.
Parsedown extra replaces every occurance of "placeholder" with well … something else. This is of course some very unexpected and unwanted behavior. This patch uses a random sha1 hash instead of "placeholder", because it is very likely that you use the latter one sooner or later in your HTML code, e.g. the name of a css class or a filename.

Someone already created an issue for that in the parsedown extra repository, but it does not seem, that this is fixed too soon. (https://github.com/erusev/parsedown-extra/issues/63).