Added thumb options to preserve icc color profile and/or meta data.
To use one of the two options, add ont or both of the following lines to config: `c::set('thumbs.icc.profile', true);` and/or `c::set('thumbs.meta.data', true);`. The meta data setting trumps the icc profile. When meta data is set to `true`, the color (icc) profile is preserved no matter if the icc profile setting is set to `true` or `false`. When both settings are set to `true`, only the `meta` flag is added to the thumb file name.