Fix #581: prevent infinite redirect on lang detect
Since this part only applies to `language.detect=true` on the home page of the site, we have to add a `/` before the `?` of the added query string. Otherwise the if-statement part  `trim(url::current(), '/') !== rtrim($url, ‘/‘)`(from line 264) would be `true` since `url::current()` provides the url with the `/` while `$url` has not contained this `/` so far.